backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
pyttsx3==2.90
SpeechRecognition==3.10.1
pyautogui==0.9.54
psutil==5.9.6
requests==2.31.0
google-generativeai>=0.8.0
comtypes
